Posibilidad de crear cursos, mediante el servlet, dao y servicios de Cursos, modificacion de la pantalla del administrador y del superadministrador, se ha tenido que modificar el pojo de usuario para la pantalla del superadministrador, añadida una query al archivo de constantes.

+++ ejercicioBaseDatos/src/main/java/model/Alumno.java
@@ -6,9 +6,9 @@ package model;
  */
 public class Alumno {
 
-    private int id;
+    private long id;
     private String nombre;
-
+    
     public Alumno() {
     }
 
@@ -17,11 +17,11 @@ public class Alumno {
         this.nombre = nombre;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
